orchestration/order-orchestration/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
httpx==0.27.0
aiokafka==0.11.0
redis[asyncio]==5.0.4
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
